+++ Code/Genrpg.Shared/Inventory/PlayerData/Item.cs
@@ -20,7 +20,7 @@ namespace Genrpg.Shared.Inventory.PlayerData
         [Key(9)] public long EquipSlotId { get; set; }
         [Key(10)] public long BuyCost { get; set; }
         [Key(11)] public long SellValue { get; set; }
-        [Key(12)] public int Level { get; set; }
+        [Key(12)] public long Level { get; set; }
 
         [Key(13)] public List<ItemProc> Procs { get; set; } = new List<ItemProc>();
