+++ Code/Genrpg.Shared/PlayerGroups/Entities/GroupMember.cs
@@ -10,7 +10,7 @@ namespace Genrpg.Shared.PlayerGroups.Entities
     {
         [Key(0)] public string Id { get; set; }
         [Key(1)] public string Name { get; set; }
-        [Key(2)] public int Level { get; set; }
+        [Key(2)] public long Level { get; set; }
         [Key(3)] public int Rank { get; set; }
     }
 }
